honeybee: Move all of sign message rendering code to signmsg module.
