+++ bin/honeybee/src/signmsg.rs
@@ -12,10 +12,12 @@
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 // GNU General Public License for more details.
 //
-use crate::error::Result;
-use crate::multi::ColorCtx;
+use crate::error::{Error, Result};
+use crate::font::{Font, Graphic};
+use crate::multi::{ColorClassic, ColorCtx, ColorScheme, LineJustification,
+                   PageJustification, Rectangle};
 use crate::render::{PageSplitter, State};
-use crate::resource::{MsgData, SignConfig, SignMessage};
+use crate::resource::{make_name, make_tmp_name};
 use gift::Encoder;
 use gift::block::{
     Application, ColorTableConfig, ColorTableExistence, ColorTableOrdering,
@@ -23,13 +25,434 @@ use gift::block::{
     LogicalScreenDesc, Preamble,
 };
 use pix::{Gray8, Palette, Raster, RasterBuilder, Rgb8};
-use std::io::Write;
+use postgres::{Connection, rows::Row};
+use std::collections::{HashMap, HashSet};
+use std::convert::TryInto;
+use std::fs::{File, rename, remove_file, read_dir};
+use std::io::{BufReader, BufWriter, Write};
+use std::path::{Path, PathBuf};
+use std::time::Instant;
 
 /// Maximum pixel width of DMS images
 const PIX_WIDTH: f32 = 450.0;
 
 /// Maximum pixel height of DMS images
 const PIX_HEIGHT: f32 = 100.0;
+
+/// DMS attribute
+#[derive(Serialize, Deserialize)]
+struct DmsAttribute {
+    name  : String,
+    value : String,
+}
+
+/// Sign configuration
+#[derive(Serialize, Deserialize)]
+struct SignConfig {
+    name        : String,
+    face_width  : i32,
+    face_height : i32,
+    border_horiz: i32,
+    border_vert : i32,
+    pitch_horiz : i32,
+    pitch_vert  : i32,
+    pixel_width : i32,
+    pixel_height: i32,
+    char_width  : i32,
+    char_height : i32,
+    monochrome_foreground: i32,
+    monochrome_background: i32,
+    color_scheme: String,
+    default_font: Option<String>,
+}
+
+/// Sign detail
+#[derive(Serialize, Deserialize)]
+struct SignDetail {
+    name           : String,
+    dms_type       : String,
+    portable       : bool,
+    technology     : String,
+    sign_access    : String,
+    legend         : String,
+    beacon_type    : String,
+    hardware_make  : String,
+    hardware_model : String,
+    software_make  : String,
+    software_model : String,
+    supported_tags : i32,
+    max_pages      : i32,
+    max_multi_len  : i32,
+}
+
+/// Data needed for rendering sign messages
+struct MsgData {
+    attrs   : HashMap<String, DmsAttribute>,
+    configs : HashMap<String, SignConfig>,
+    fonts   : HashMap<i32, Font>,
+    graphics: HashMap<i32, Graphic>,
+    gifs    : HashSet<PathBuf>,
+}
+
+/// Sign message
+#[derive(Serialize)]
+struct SignMessage {
+    name          : String,
+    sign_config   : String,
+    incident      : Option<String>,
+    multi         : String,
+    beacon_enabled: bool,
+    prefix_page   : bool,
+    msg_priority  : i32,
+    sources       : String,
+    owner         : Option<String>,
+    duration      : Option<i32>,
+}
+
+impl DmsAttribute {
+    /// Load DMS attributes from a JSON file
+    fn load(dir: &Path) -> Result<HashMap<String, DmsAttribute>> {
+        debug!("DmsAttribute::load");
+        let mut n = PathBuf::new();
+        n.push(dir);
+        n.push("dms_attribute");
+        let r = BufReader::new(File::open(&n)?);
+        let mut attrs = HashMap::new();
+        let j: Vec<DmsAttribute> = serde_json::from_reader(r)?;
+        for da in j {
+            let an = da.name.clone();
+            attrs.insert(an, da);
+        }
+        Ok(attrs)
+    }
+}
+
+impl SignConfig {
+    /// Load sign configurations from a JSON file
+    fn load(dir: &Path) -> Result<HashMap<String, SignConfig>> {
+        debug!("SignConfig::load");
+        let mut n = PathBuf::new();
+        n.push(dir);
+        n.push("sign_config");
+        let r = BufReader::new(File::open(&n)?);
+        let mut configs = HashMap::new();
+        let j: Vec<SignConfig> = serde_json::from_reader(r)?;
+        for c in j {
+            let cn = c.name.clone();
+            configs.insert(cn, c);
+        }
+        Ok(configs)
+    }
+    /// Get config name
+    fn name(&self) -> &str {
+        &self.name
+    }
+    /// Get default font
+    fn default_font(&self) -> Option<&str> {
+        self.default_font.as_ref().map(String::as_str)
+    }
+    /// Get the horizontal border (mm).
+    /// Sanity check included in case the sign vendor supplies stupid values.
+    fn horizontal_border_mm(&self) -> f32 {
+        let excess_mm = self.horizontal_excess_mm();
+        let border_horiz = self.border_horiz as f32;
+        border_horiz.min(0_f32.max(excess_mm / 2.0))
+    }
+    /// Get the face width (mm)
+    fn face_width(&self) -> f32 {
+        self.face_width as f32
+    }
+    /// Get the face height (mm)
+    fn face_height(&self) -> f32 {
+        self.face_height as f32
+    }
+    /// Get the horizontal excess (mm).
+    fn horizontal_excess_mm(&self) -> f32 {
+        let pixels_mm = (self.pitch_horiz * (self.pixel_width - 1)) as f32;
+        self.face_width() - pixels_mm
+    }
+    /// Get the character height
+    fn char_height(&self) -> Result<u8> {
+        Ok(self.char_height.try_into()?)
+    }
+    /// Get the character width
+    fn char_width(&self) -> Result<u8> {
+        Ok(self.char_width.try_into()?)
+    }
+    /// Get the horizontal character offset (mm)
+    fn char_offset_mm(&self, x: u32) -> f32 {
+        if self.char_width > 1 {
+            let gap = (x / self.char_width as u32) as f32;
+            gap * self.char_gap_mm()
+        } else {
+            0.0
+        }
+    }
+    /// Get the character gap (mm)
+    fn char_gap_mm(&self) -> f32 {
+        let excess_mm = self.horizontal_excess_mm();
+        let border_mm = self.horizontal_border_mm() * 2.0;
+        let gaps = self.char_gaps() as f32;
+        if excess_mm > border_mm && gaps > 0.0 {
+            (excess_mm - border_mm) / gaps
+        } else {
+            0.0
+        }
+    }
+    /// Get the number of gaps between characters
+    fn char_gaps(&self) -> i32 {
+        if self.char_width > 1 && self.pixel_width > self.char_width {
+            (self.pixel_width / self.char_width) - 1
+        } else {
+            0
+        }
+    }
+    /// Get the X-position of a pixel on the sign (from 0 to 1)
+    fn pixel_x(&self, x: u32) -> f32 {
+        let hb = self.horizontal_border_mm();
+        let co = self.char_offset_mm(x);
+        let pos = hb + co + (self.pitch_horiz * x as i32) as f32;
+        pos / self.face_width()
+    }
+    /// Get the vertical border (mm).
+    /// Sanity check included in case the sign vendor supplies stupid values.
+    fn vertical_border_mm(&self) -> f32 {
+        let excess_mm = self.vertical_excess_mm();
+        let border_vert = self.border_vert as f32;
+        border_vert.min(0_f32.max(excess_mm / 2.0))
+    }
+    /// Get the vertical excess (mm).
+    fn vertical_excess_mm(&self) -> f32 {
+        let pixels_mm = (self.pitch_vert * (self.pixel_height - 1)) as f32;
+        self.face_height() - pixels_mm
+    }
+    /// Get the number of gaps between lines
+    fn line_gaps(&self) -> i32 {
+        if self.char_height > 1 && self.pixel_height > self.char_height {
+            (self.pixel_height / self.char_height) - 1
+        } else {
+            0
+        }
+    }
+    /// Get the vertical line offset (mm)
+    fn line_offset_mm(&self, y: u32) -> f32 {
+        if self.char_height > 1 {
+            let gap = (y / self.char_height as u32) as f32;
+            gap * self.line_gap_mm()
+        } else {
+            0.0
+        }
+    }
+    /// Get the line gap (mm)
+    fn line_gap_mm(&self) -> f32 {
+        let excess_mm = self.vertical_excess_mm();
+        let border_mm = self.vertical_border_mm() * 2.0;
+        let gaps = self.line_gaps() as f32;
+        if excess_mm > border_mm && gaps > 0.0 {
+            (excess_mm - border_mm) / gaps
+        } else {
+            0.0
+        }
+    }
+    /// Get the Y-position of a pixel on the sign (from 0 to 1)
+    fn pixel_y(&self, y: u32) -> f32 {
+        let vb = self.vertical_border_mm();
+        let lo = self.line_offset_mm(y);
+        let pos = vb + lo + (self.pitch_vert * y as i32) as f32;
+        pos / self.face_height()
+    }
+    /// Get the color scheme value
+    fn color_scheme(&self) -> ColorScheme {
+        self.color_scheme[..].into()
+    }
+    /// Get the default foreground color
+    fn foreground_default(&self) -> i32 {
+        match self.color_scheme() {
+            ColorScheme::ColorClassic |
+            ColorScheme::Color24Bit => ColorClassic::Amber.rgb(),
+            _ => self.monochrome_foreground,
+        }
+    }
+    /// Get the default background color
+    fn background_default(&self) -> i32 {
+        match self.color_scheme() {
+            ColorScheme::ColorClassic |
+            ColorScheme::Color24Bit => ColorClassic::Black.rgb(),
+            _ => self.monochrome_background,
+        }
+    }
+    /// Get the default text rectangle
+    fn text_rect_default(&self) -> Result<Rectangle> {
+        let w = self.pixel_width.try_into()?;
+        let h = self.pixel_height.try_into()?;
+        Ok(Rectangle::new(1, 1, w, h))
+    }
+}
+
+impl MsgData {
+    /// Load message data from a file path
+    fn load(dir: &Path) -> Result<Self> {
+        debug!("MsgData::load");
+        let attrs = DmsAttribute::load(dir)?;
+        let configs = SignConfig::load(dir)?;
+        let fonts = Font::load(dir)?;
+        let graphics = Graphic::load(dir)?;
+        let gifs = gif_listing(dir)?;
+        Ok(MsgData {
+            attrs,
+            configs,
+            fonts,
+            graphics,
+            gifs,
+        })
+    }
+    /// Lookup a config
+    fn config(&self, s: &SignMessage) -> Result<&SignConfig> {
+        let cfg = &s.sign_config;
+        match self.configs.get(cfg) {
+            Some(c) => Ok(c),
+            None => Err(Error::UnknownResource(format!("Config: {}", cfg))),
+        }
+    }
+    /// Check if a .gif file is in the listing.
+    ///
+    /// Returns true if the file exists.
+    fn check_gif_listing(&mut self, n: &PathBuf) -> bool {
+        self.gifs.remove(n)
+    }
+    /// Delete out-of-date .gif files
+    fn delete_gifs(&mut self) -> Result<()> {
+        for p in self.gifs.drain() {
+            info!("delete gif: {:?}", &p);
+            if let Err(e) = remove_file(&p) {
+                error!("{:?}", e);
+            }
+        }
+        Ok(())
+    }
+    /// Get an attribute (seconds) in u8 deciseconds
+    fn get_as_ds(&self, name: &str) -> Option<u8> {
+        if let Some(attr) = self.attrs.get(name) {
+            if let Ok(sec) = attr.name.parse::<f32>() {
+                let ds = sec * 10.0;
+                if ds >= 0.0 && ds <= 255.0 {
+                    return Some(ds as u8);
+                }
+            }
+        }
+        debug!("MsgData::get_as_ds, {} missing!", name);
+        None
+    }
+    /// Get the default page on time
+    fn page_on_default_ds(&self) -> u8 {
+        self.get_as_ds("dms_page_on_default_secs").unwrap_or(20)
+    }
+    /// Get the default page off time
+    fn page_off_default_ds(&self) -> u8 {
+        self.get_as_ds("dms_page_off_default_secs").unwrap_or(0)
+    }
+    /// Get the default page justification
+    fn page_justification_default(&self) -> PageJustification {
+        const ATTR: &str = "dms_default_justification_page";
+        if let Some(attr) = self.attrs.get(ATTR) {
+            if let Some(pj) = PageJustification::new(&attr.name) {
+                return pj;
+            }
+        }
+        debug!("MsgData::page_justification_default, {} missing!", ATTR);
+        PageJustification::Top
+    }
+    /// Get the default line justification
+    fn line_justification_default(&self) -> LineJustification {
+        const ATTR: &str = "dms_default_justification_line";
+        if let Some(attr) = self.attrs.get(ATTR) {
+            if let Some(lj) = LineJustification::new(&attr.name) {
+                return lj;
+            }
+        }
+        debug!("MsgData::line_justification_default, {} missing!", ATTR);
+        LineJustification::Center
+    }
+    /// Get the default font number
+    fn font_default(&self, fname: Option<&str>) -> Result<u8> {
+        match fname {
+            Some(fname) => {
+                match self.fonts.values().find(|f| &f.name == fname) {
+                    Some(font) => Ok(font.f_number.try_into()?),
+                    None => {
+                        Err(Error::UnknownResource(format!("Font: {}", fname)))
+                    },
+                }
+            },
+            None => Ok(1),
+        }
+    }
+    /// Get font mapping
+    fn fonts(&self) -> &HashMap<i32, Font> {
+        &self.fonts
+    }
+    /// Get graphic mapping
+    fn graphics(&self) -> &HashMap<i32, Graphic> {
+        &self.graphics
+    }
+}
+
+impl SignMessage {
+    /// Get the SQL to query all sign messages
+    fn sql() -> &'static str {
+        "SELECT name, sign_config, incident, multi, beacon_enabled, \
+                prefix_page, msg_priority, sources, owner, duration \
+        FROM sign_message_view \
+        ORDER BY name"
+    }
+    /// Produce a sign message from one Row
+    fn from_row(row: &Row) -> Self {
+        SignMessage {
+            name          : row.get(0),
+            sign_config   : row.get(1),
+            incident      : row.get(2),
+            multi         : row.get(3),
+            beacon_enabled: row.get(4),
+            prefix_page   : row.get(5),
+            msg_priority  : row.get(6),
+            sources       : row.get(7),
+            owner         : row.get(8),
+            duration      : row.get(9),
+        }
+    }
+    /// Get the MULTI string
+    fn multi(&self) -> &str {
+        &self.multi
+    }
+    /// Render into a .gif file
+    fn render<W: Write>(&self, msg_data: &MsgData, writer: W) -> Result<()> {
+        let cfg = msg_data.config(self)?;
+        let (preamble, frames) = render_sign_msg_cfg(self, msg_data, cfg)?;
+        write_gif(writer, preamble, frames)
+    }
+}
+
+/// Lookup a listing of gif files
+fn gif_listing(dir: &Path) -> Result<HashSet<PathBuf>> {
+    let mut gifs = HashSet::new();
+    let mut img = PathBuf::new();
+    img.push(dir);
+    img.push("img");
+    if img.is_dir() {
+        for f in read_dir(img)? {
+            let f = f?;
+            if f.file_type()?.is_file() {
+                let p = f.path();
+                let b = if let Some(ext) = p.extension() { ext == "gif" }
+                        else { false };
+                if b {
+                    gifs.insert(p);
+                }
+            }
+        }
+    }
+    Ok(gifs)
+}
 
 /// Calculate the size of rendered DMS
 fn calculate_size(cfg: &SignConfig) -> Result<(u16, u16)> {
@@ -136,15 +559,6 @@ fn render_circle(raster: &mut Raster<Gray8>, palette: &mut Palette<Rgb8>,
             }
         }
     }
-}
-
-/// Render a sign message into a .gif file
-pub fn render<W: Write>(s: &SignMessage, msg_data: &MsgData, writer: W)
-    -> Result<()>
-{
-    let cfg = msg_data.config(s)?;
-    let (preamble, frames) = render_sign_msg_cfg(s, msg_data, cfg)?;
-    write_gif(writer, preamble, frames)
 }
 
 /// Write a .gif file
@@ -270,4 +684,55 @@ fn render_state_default(msg_data: &MsgData, cfg: &SignConfig) -> Result<State> {
                   just_line,
                   font,
     ))
+}
+
+/// Query the sign messages.
+///
+/// * `conn` The database connection.
+/// * `w` Writer for the file.
+/// * `dir` Output file directory.
+pub fn query_sign_msg<W: Write>(conn: &Connection, mut w: W, dir: &Path)
+    -> Result<u32>
+{
+    let mut msg_data = MsgData::load(dir)?;
+    let mut c = 0;
+    w.write("[".as_bytes())?;
+    for row in &conn.query(SignMessage::sql(), &[])? {
+        if c > 0 { w.write(",".as_bytes())?; }
+        w.write("\n".as_bytes())?;
+        let s = SignMessage::from_row(&row);
+        w.write(serde_json::to_string(&s)?.as_bytes())?;
+        fetch_sign_msg(&s, dir, &mut msg_data)?;
+        c += 1;
+    }
+    w.write("]\n".as_bytes())?;
+    msg_data.delete_gifs()?;
+    Ok(c)
+}
+
+/// Check and fetch one sign message (into a .gif file).
+fn fetch_sign_msg(s: &SignMessage, dir: &Path, msg_data: &mut MsgData)
+    -> Result<()>
+{
+    let mut img = PathBuf::new();
+    img.push(dir);
+    img.push("img");
+    let mut g = String::new();
+    g.push_str(&s.name);
+    g.push_str(&".gif");
+    let n = make_name(&img.as_path(), &g);
+    debug!("fetch: {:?}", n);
+    if !msg_data.check_gif_listing(&n) {
+        let tn = make_tmp_name(&img.as_path(), &g);
+        let writer = BufWriter::new(File::create(&tn)?);
+        let t = Instant::now();
+        if let Err(e) = s.render(msg_data, writer) {
+            warn!("{},cfg={},multi={} {:?}", &s.name, s.sign_config, s.multi,e);
+            remove_file(&tn)?;
+            return Ok(());
+        };
+        rename(tn, &n)?;
+        info!("{}.gif rendered in {:?}", &s.name, t.elapsed());
+    }
+    Ok(())
 }
